requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.1
openai==1.30.1
apscheduler==3.10.4
python-dotenv==1.0.1
httpx==0.27.0
python-telegram-bot==21.3
python-multipart==0.0.9
Pillow==10.3.0
